requirements.txt
sqlalchemy==1.4.36
fastapi==0.78.0
psycopg2-binary==2.9.3
python-jose==3.3.0
passlib[bcrypt]==1.7.4